# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "This is a REST API application made with Express",
        "version": "1.0.0",
        "title": "Express API for Users and Forumposts",
        "contact": {
            "name": "The repo is found here:",
            "url": "https://github.com/sararazm/gruppsarbete-backend"
        }
    },
        "schemes": ["http"],
        "host": "localhost:3030",
        "basePath": "/api", 
        "paths": {
                "/user" : {
                    "get": {
                        "summary": "Get all users",
                        "desription": "Get all users",
                        "produces": ["appliction/json"],
                        "parameters": [],
                        "responses": {
                            "200": {
                                "description": "Successully fetched all users",
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/usersResponse"
                                    }
                                } 
                            },
                            "400": {
                                "description": "Invalid",
                                "schema": {
                                    "$ref": "#/definitions/InvalidResponse"
                                }
                            }
                        }
                    },
                
                    "post": {
                        "summary": "Create a user",
                        "description": "Create a user",
                        "produces": ["application/json"],
                        "consumes": ["application/json"],
                        "parameters": [{
                            "in":"body",
                            "name": "body",
                            "description": "User object",
                            "required": true,
                            "schema": {
                                "type": "object",
                                "properties": {
                                    
                                }
                            }

                        }]



                    }
                }
        }
    }